move dup opcodes
